requirements.txt
fastapi
uvicorn
langchain
langchain-ollama
langgraph
sentence-transformers
pinecone-client
pydantic
python-multipart
python-dotenv
PyGithub

